app/worker/requirements.txt
google-cloud-pubsub==2.20.2
google-cloud-secret-manager==2.20.2
